refactor view component names for improved consistency
