+++ Modules/feature-dashboard/Sources/UI/Dashboard/Tab/Home/HomeTabView.swift
@@ -87,7 +87,7 @@ private func content(
 ) -> some View {
   ScrollView {
     VStack(alignment: .leading, spacing: SPACING_MEDIUM) {
-      ContentHeader(
+      ContentHeaderView(
         config: viewState.contentHeaderConfig
       )
