+++ Modules/feature-issuance/Sources/UI/Document/Code/OfferCodeView.swift
@@ -82,7 +82,7 @@ private func content(
   codeIsFocused: Binding<Bool>
 ) -> some View {
 
-  ContentHeader(
+  ContentHeaderView(
     config: viewState.contentHeaderConfig
   )
