+++ Modules/logic-ui/Sources/DesignSystem/Component/List/WrapExpandableListView.swift
@@ -70,7 +70,7 @@ public struct WrapExpandableListView<T: Sendable>: View {
       expandableItemView(item)
 
       if index < items.count - 1 {
-        ListDivider()
+        ListDividerView()
       }
     }
   }
