+++ Modules/logic-ui/Sources/DesignSystem/Component/List/WrapListItemsView.swift
@@ -47,7 +47,7 @@ public struct WrapListItemsView: View {
         }
       )
       if index < listItems.count - 1 {
-        ListDivider()
+        ListDividerView()
       }
     }
   }
